+++ zcash_client_backend/src/wallet.rs
@@ -7,12 +7,10 @@ use zcash_primitives::{
     keys::OutgoingViewingKey,
     legacy::TransparentAddress,
     merkle_tree::IncrementalWitness,
-    sapling::{
-        note::ExtractedNoteCommitment, Diversifier, Node, Note, Nullifier, PaymentAddress, Rseed,
-    },
+    sapling,
     transaction::{
         components::{
-            sapling,
+            sapling::fees as sapling_fees,
             transparent::{self, OutPoint, TxOut},
             Amount,
         },
@@ -27,10 +25,8 @@ use zcash_primitives::{
 pub struct WalletTx<N> {
     pub txid: TxId,
     pub index: usize,
-    pub num_spends: usize,
-    pub num_outputs: usize,
-    pub shielded_spends: Vec<WalletShieldedSpend>,
-    pub shielded_outputs: Vec<WalletShieldedOutput<N>>,
+    pub sapling_spends: Vec<WalletSaplingSpend>,
+    pub sapling_outputs: Vec<WalletSaplingOutput<N>>,
 }
 
 #[derive(Debug, Clone)]
@@ -90,24 +86,24 @@ impl transparent::fees::InputView for WalletTransparentOutput {
 /// A subset of a [`SpendDescription`] relevant to wallets and light clients.
 ///
 /// [`SpendDescription`]: zcash_primitives::transaction::components::SpendDescription
-pub struct WalletShieldedSpend {
+pub struct WalletSaplingSpend {
     pub index: usize,
-    pub nf: Nullifier,
+    pub nf: sapling::Nullifier,
     pub account: AccountId,
 }
 
 /// A subset of an [`OutputDescription`] relevant to wallets and light clients.
 ///
 /// [`OutputDescription`]: zcash_primitives::transaction::components::OutputDescription
-pub struct WalletShieldedOutput<N> {
+pub struct WalletSaplingOutput<N> {
     pub index: usize,
-    pub cmu: ExtractedNoteCommitment,
+    pub cmu: sapling::note::ExtractedNoteCommitment,
     pub ephemeral_key: EphemeralKeyBytes,
     pub account: AccountId,
-    pub note: Note,
-    pub to: PaymentAddress,
+    pub note: sapling::Note,
+    pub to: sapling::PaymentAddress,
     pub is_change: bool,
-    pub witness: IncrementalWitness<Node>,
+    pub witness: IncrementalWitness<sapling::Node>,
     pub nf: N,
 }
 
@@ -115,13 +111,13 @@ pub struct WalletShieldedOutput<N> {
 /// with sufficient information for use in note selection.
 pub struct SpendableNote<NoteRef> {
     pub note_id: NoteRef,
-    pub diversifier: Diversifier,
+    pub diversifier: sapling::Diversifier,
     pub note_value: Amount,
-    pub rseed: Rseed,
-    pub witness: IncrementalWitness<Node>,
+    pub rseed: sapling::Rseed,
+    pub witness: IncrementalWitness<sapling::Node>,
 }
 
-impl<NoteRef> sapling::fees::InputView<NoteRef> for SpendableNote<NoteRef> {
+impl<NoteRef> sapling_fees::InputView<NoteRef> for SpendableNote<NoteRef> {
     fn note_id(&self) -> &NoteRef {
         &self.note_id
     }