+++ zcash_client_backend/src/welding_rig.rs
@@ -20,7 +20,7 @@ use zcash_primitives::{
 use crate::{
     proto::compact_formats::CompactBlock,
     scan::{Batch, BatchRunner, Tasks},
-    wallet::{WalletShieldedOutput, WalletShieldedSpend, WalletTx},
+    wallet::{WalletSaplingOutput, WalletSaplingSpend, WalletTx},
 };
 
 /// A key that can be used to perform trial decryption and nullifier
@@ -115,7 +115,7 @@ impl ScanningKey for SaplingIvk {
 ///
 /// Returns a vector of [`WalletTx`]s belonging to any of the given
 /// [`ScanningKey`]s. If scanning with a full viewing key, the nullifiers
-/// of the resulting [`WalletShieldedOutput`]s will also be computed.
+/// of the resulting [`WalletSaplingOutput`]s will also be computed.
 ///
 /// The given [`CommitmentTree`] and existing [`IncrementalWitness`]es are
 /// incremented appropriately.
@@ -123,7 +123,7 @@ impl ScanningKey for SaplingIvk {
 /// The implementation of [`ScanningKey`] may either support or omit the computation of
 /// the nullifiers for received notes; the implementation for [`ExtendedFullViewingKey`]
 /// will derive the nullifiers for received notes and return them as part of the resulting
-/// [`WalletShieldedOutput`]s, whereas the implementation for [`SaplingIvk`] cannot
+/// [`WalletSaplingOutput`]s, whereas the implementation for [`SaplingIvk`] cannot
 /// do so and will return the unit value in those outputs instead.
 ///
 /// [`ExtendedFullViewingKey`]: zcash_primitives::zip32::ExtendedFullViewingKey
@@ -132,7 +132,7 @@ impl ScanningKey for SaplingIvk {
 /// [`ScanningKey`]: crate::welding_rig::ScanningKey
 /// [`CommitmentTree`]: zcash_primitives::merkle_tree::CommitmentTree
 /// [`IncrementalWitness`]: zcash_primitives::merkle_tree::IncrementalWitness
-/// [`WalletShieldedOutput`]: crate::wallet::WalletShieldedOutput
+/// [`WalletSaplingOutput`]: crate::wallet::WalletSaplingOutput
 /// [`WalletTx`]: crate::wallet::WalletTx
 pub fn scan_block<P: consensus::Parameters + Send + 'static, K: ScanningKey>(
     params: &P,
@@ -211,8 +211,6 @@ pub(crate) fn scan_block_with_runner<
     for tx in block.vtx.into_iter() {
         let txid = tx.txid();
         let index = tx.index as usize;
-        let num_spends = tx.spends.len();
-        let num_outputs = tx.outputs.len();
 
         // Check for spent notes
         // The only step that is not constant-time is the filter() at the end.
@@ -233,7 +231,7 @@ pub(crate) fn scan_block_with_runner<
                         CtOption::new(AccountId::from(0), 0.into()),
                         |first, next| CtOption::conditional_select(&next, &first, first.is_some()),
                     )
-                    .map(|account| WalletShieldedSpend {
+                    .map(|account| WalletSaplingSpend {
                         index,
                         nf: spend_nf,
                         account,
@@ -248,7 +246,7 @@ pub(crate) fn scan_block_with_runner<
             shielded_spends.iter().map(|spend| spend.account).collect();
 
         // Check for incoming notes while incrementing tree and witnesses
-        let mut shielded_outputs: Vec<WalletShieldedOutput<K::Nf>> = vec![];
+        let mut shielded_outputs: Vec<WalletSaplingOutput<K::Nf>> = vec![];
         {
             // Grab mutable references to new witnesses from previous transactions
             // in this block so that we can update them. Scoped so we don't hold
@@ -256,7 +254,7 @@ pub(crate) fn scan_block_with_runner<
             let mut block_witnesses: Vec<_> = wtxs
                 .iter_mut()
                 .flat_map(|tx| {
-                    tx.shielded_outputs
+                    tx.sapling_outputs
                         .iter_mut()
                         .map(|output| &mut output.witness)
                 })
@@ -357,7 +355,7 @@ pub(crate) fn scan_block_with_runner<
                     let witness = IncrementalWitness::from_tree(tree);
                     let nf = K::sapling_nf(&nk, &note, &witness);
 
-                    shielded_outputs.push(WalletShieldedOutput {
+                    shielded_outputs.push(WalletSaplingOutput {
                         index,
                         cmu: output.cmu,
                         ephemeral_key: output.ephemeral_key.clone(),
@@ -376,10 +374,8 @@ pub(crate) fn scan_block_with_runner<
             wtxs.push(WalletTx {
                 txid,
                 index,
-                num_spends,
-                num_outputs,
-                shielded_spends,
-                shielded_outputs,
+                sapling_spends: shielded_spends,
+                sapling_outputs: shielded_outputs,
             });
         }
     }
@@ -569,16 +565,14 @@ mod tests {
 
             let tx = &txs[0];
             assert_eq!(tx.index, 1);
-            assert_eq!(tx.num_spends, 1);
-            assert_eq!(tx.num_outputs, 1);
-            assert_eq!(tx.shielded_spends.len(), 0);
-            assert_eq!(tx.shielded_outputs.len(), 1);
-            assert_eq!(tx.shielded_outputs[0].index, 0);
-            assert_eq!(tx.shielded_outputs[0].account, account);
-            assert_eq!(tx.shielded_outputs[0].note.value().inner(), 5);
+            assert_eq!(tx.sapling_spends.len(), 0);
+            assert_eq!(tx.sapling_outputs.len(), 1);
+            assert_eq!(tx.sapling_outputs[0].index, 0);
+            assert_eq!(tx.sapling_outputs[0].account, account);
+            assert_eq!(tx.sapling_outputs[0].note.value().inner(), 5);
 
             // Check that the witness root matches
-            assert_eq!(tx.shielded_outputs[0].witness.root(), tree.root());
+            assert_eq!(tx.sapling_outputs[0].witness.root(), tree.root());
         }
 
         go(false);
@@ -632,16 +626,14 @@ mod tests {
 
             let tx = &txs[0];
             assert_eq!(tx.index, 1);
-            assert_eq!(tx.num_spends, 1);
-            assert_eq!(tx.num_outputs, 1);
-            assert_eq!(tx.shielded_spends.len(), 0);
-            assert_eq!(tx.shielded_outputs.len(), 1);
-            assert_eq!(tx.shielded_outputs[0].index, 0);
-            assert_eq!(tx.shielded_outputs[0].account, AccountId::from(0));
-            assert_eq!(tx.shielded_outputs[0].note.value().inner(), 5);
+            assert_eq!(tx.sapling_spends.len(), 0);
+            assert_eq!(tx.sapling_outputs.len(), 1);
+            assert_eq!(tx.sapling_outputs[0].index, 0);
+            assert_eq!(tx.sapling_outputs[0].account, AccountId::from(0));
+            assert_eq!(tx.sapling_outputs[0].note.value().inner(), 5);
 
             // Check that the witness root matches
-            assert_eq!(tx.shielded_outputs[0].witness.root(), tree.root());
+            assert_eq!(tx.sapling_outputs[0].witness.root(), tree.root());
         }
 
         go(false);
@@ -672,12 +664,10 @@ mod tests {
 
         let tx = &txs[0];
         assert_eq!(tx.index, 1);
-        assert_eq!(tx.num_spends, 1);
-        assert_eq!(tx.num_outputs, 1);
-        assert_eq!(tx.shielded_spends.len(), 1);
-        assert_eq!(tx.shielded_outputs.len(), 0);
-        assert_eq!(tx.shielded_spends[0].index, 0);
-        assert_eq!(tx.shielded_spends[0].nf, nf);
-        assert_eq!(tx.shielded_spends[0].account, account);
+        assert_eq!(tx.sapling_spends.len(), 1);
+        assert_eq!(tx.sapling_outputs.len(), 0);
+        assert_eq!(tx.sapling_spends[0].index, 0);
+        assert_eq!(tx.sapling_spends[0].nf, nf);
+        assert_eq!(tx.sapling_spends[0].account, account);
     }
 }